+++ src/lib/components/AIMessage.svelte
