+++ src/lib/components/AutoRecordIcon.svelte
