+++ src/lib/components/DouyinIcon.svelte
