+++ src/lib/components/SidebarItem.svelte
