+++ src/lib/components/ToolMessage.svelte
